requirements-simulator.txt
# Dépendances pour collar_simulator.py (tests MQTT)
paho-mqtt>=2.0.0
